feat: Modify ORM entities and relationships (#10)
