+++ module-exchange/src/main/java/com/example/module_exchange/exchange/TransactionCategory.java
@@ -1,0 +1,5 @@
+package com.example.module_exchange.exchange;
+
+public enum TransactionCategory {
+    BASIC ,STOCK_TRADE, EXCHANGE
+}
