+++ module-exchange/src/main/java/com/example/module_exchange/exchange/exchangeHistory/ExchangeHistory.java
@@ -1,7 +1,10 @@
-package com.example.module_exchange.exchange;
+package com.example.module_exchange.exchange.exchangeHistory;
 
+import com.example.module_exchange.exchange.exchangeCurrency.ExchangeCurrency;
 import com.example.module_utility.entity.BaseEntity;
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -9,7 +12,9 @@ import java.math.BigDecimal;
 
 @Entity
 @Getter
+@Builder
 @NoArgsConstructor
+@AllArgsConstructor
 public class ExchangeHistory extends BaseEntity {
 
     @Id
@@ -22,8 +27,7 @@ public class ExchangeHistory extends BaseEntity {
     private BigDecimal exchangeRate;
 
     @ManyToOne
-    @JoinColumn(name = "wallet_id", nullable = false)
-    ExchangeWallet exchangeWallet;
+    @JoinColumn(name="exchange_currency_id",nullable = false)
+    private ExchangeCurrency exchangeCurrency;
 
 }
-
