+++ module-exchange/src/main/java/com/example/module_exchange/exchange/stockTradeHistory/TradeType.java
@@ -1,0 +1,5 @@
+package com.example.module_exchange.exchange.stockTradeHistory;
+
+public enum TradeType {
+    BUY, SELL
+}
