+++ module-exchange/src/main/java/com/example/module_exchange/exchange/transactionHistory/TransactionType.java
@@ -1,0 +1,5 @@
+package com.example.module_exchange.exchange.transactionHistory;
+
+public enum TransactionType {
+    DEPOSIT, WITHDRAWAL
+}
