--- module-trip/src/main/java/com/example/module_trip/account/TradeType.java
@@ -1,5 +1,0 @@
-package com.example.module_trip.account;
-
-public enum TradeType {
-    BUY,SELL
-}
