--- module-trip/src/main/java/com/example/module_trip/account/TransactionType.java
@@ -1,5 +1,0 @@
-package com.example.module_trip.account;
-
-public enum TransactionType {
-    DEPOSIT, WITHDRAWAL
-}
